requirements.txt
pandas
python-dotenv
fastapi
uvicorn
openai>=1.40.0
